Rename calendarApp to snake case
