+++ python_anywhere/calendar_app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CalendarappConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'calendarApp'
+    name = 'calendar_app'
